Add model of eath and telescope for scanning.
A surprising amount of effort went into this. The earth and telescope had to be rescaled and positioned, and the scanner view needed some offset work as well.
In addition, a second camera has been added. It is bound to Display 2, and exists for the purpose of seeing the game view without any UI.